+++ src/Hyjinx.LibHac/Tools/FsSystem/NcaUtils/NcaFsPatchInfo.cs
@@ -27,6 +27,8 @@ public struct NcaFsPatchInfo
         set => Data.RelocationTreeSize = value;
     }
 
+    #if IS_TPM_BYPASS_ENABLED
+    
     public long EncryptionTreeOffset
     {
         get => Data.EncryptionTreeOffset;
@@ -39,8 +41,13 @@ public struct NcaFsPatchInfo
         set => Data.EncryptionTreeSize = value;
     }
 
+    #endif
+    
     public Span<byte> RelocationTreeHeader => _data.Span.Slice(0x10, 0x10);
+    
+    #if IS_TPM_BYPASS_ENABLED
     public Span<byte> EncryptionTreeHeader => _data.Span.Slice(0x30, 0x10);
+    #endif
 
     [StructLayout(LayoutKind.Explicit)]
     private struct PatchInfoStruct
